persons

+++ Pkdd/pkdd-ng-client/src/app/persons/components/person-card/person-card.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Person } from '../../../models/entities/person';
 import { ServerDataStorageService } from '../../../core/services/server-data-storage.service';
+import { ConfirmService } from '../../../core/services/confirm.service';
 
 @Component({
   selector: 'pkdd-person-card',
@@ -13,13 +14,17 @@ export class PersonCardComponent implements OnInit {
   public person: Person;
 
   constructor(
-    private readonly storage: ServerDataStorageService
+    private readonly storage: ServerDataStorageService,
+    private readonly confirmer: ConfirmService
   ) { }
 
   ngOnInit() {
   }
 
-  public onDelete() {
-    this.storage.deletePerson(this.person.id);
+  public async onDelete() {
+    const shouldDelete = await this.confirmer.confirm('Удалить её навсегдааа...');
+    if (shouldDelete) {
+      this.storage.deletePerson(this.person.id);
+    }
   }
 }
